% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{create_ggplot}
\alias{create_ggplot}
\title{Create a ggplot based on prepare data}
\usage{
create_ggplot(data, ubb, labels)
}
\arguments{
\item{data}{Data}

\item{ubb}{UBB}
}
\description{
Function creates and export a plot.
}
