% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdfs.R
\name{create_pdfs}
\alias{create_pdfs}
\title{Create PDFs}
\usage{
create_pdfs(snr, audience, name, ubb, n, results, d = NULL, drop = TRUE)
}
\arguments{
\item{snr}{Schoolnumber}

\item{audience}{Reporting group}

\item{name}{Name}

\item{ubb}{UBB TRUE or FALSE}

\item{n}{Number of observations}

\item{results}{String for reporting group}

\item{d}{Duration (UBB only)}

\item{drop}{Drop the temporary files}
}
\value{
PDF file
}
\description{
Create a PDF report
}
