% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{create_PlotWeb}
\alias{create_PlotWeb}
\title{Create a Plot for the Web}
\usage{
create_PlotWeb(meta, snr, audience, report, data, ubb)
}
\arguments{
\item{meta}{Meta data}

\item{snr}{School number}

\item{audience}{Audience}

\item{report}{Report template}

\item{data}{Data}

\item{ubb}{UBB}
}
\description{
Function creates and export a plot.
}
