% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_report}
\alias{create_report}
\title{Create report}
\usage{
create_report(snr, audience, ubb, results, ganztag, stype)
}
\arguments{
\item{snr}{School number}

\item{audience}{Audience}

\item{ubb}{UBB}

\item{results}{Results: Text string for Audience}

\item{ganztag}{Ganztagsschule}

\item{stype}{School type}
}
\description{
Run all functions to create a report
}
\examples{
\dontrun{
purrr::pwalk(mylist, run_aslist)
}
}
