% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_reports}
\alias{create_reports}
\title{Create reports}
\usage{
create_reports(snr, audience, ubb, results, ganztag, stype)
}
\arguments{
\item{snr}{Schoolnumber}

\item{audience}{Report audience}

\item{ubb}{UBB TRUE or FALSE}

\item{results}{Results: Text string for audience of the report}

\item{ganztag}{Ganztagsschule TRUE or FALSE}

\item{stype}{Schooltype}
}
\description{
Run all functions to create several reports at once
}
\examples{
\dontrun{
purrr::pwalk(mylist, run_aslist)
}
}
