% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_Template}
\alias{create_Template}
\title{Generate the Rmd file for the report}
\usage{
create_Template(
  meta,
  num_bars,
  ubb,
  header,
  export = TRUE,
  file_name = "generated_document.Rmd"
)
}
\arguments{
\item{ubb}{UBB}

\item{export}{Export}

\item{file_name}{File name}

\item{x_seq}{Sequence}
}
\value{
Returns a plot
}
\description{
Functions combines the template and adds the Rmd content
from the get_rmd function
}
