% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{createWordCloud}
\alias{createWordCloud}
\title{Export a plot}
\usage{
createWordCloud(data)
}
\arguments{
\item{data}{Meta data}

\item{export}{Export}
}
\description{
Function creates and export a plot.
}
