% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debugme}
\alias{debugme}
\title{Debug the Code}
\usage{
debugme(tmp.snr, tmp.ubb, tmp.ganztag, tmp.audience, tmp.stype, tmp.results)
}
\arguments{
\item{tmp.snr}{School number}

\item{tmp.ubb}{UBB}

\item{tmp.ganztag}{Ganztagschule}

\item{tmp.audience}{Audience}

\item{tmp.stype}{School type}

\item{tmp.results}{Results string}
}
\value{
Error message
}
\description{
Function runs the code and returns error messages.
}
