% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{export_headers}
\alias{export_headers}
\title{Superseded by Parallelization: Export table headers}
\usage{
export_headers(meta, ubb)
}
\arguments{
\item{meta}{Metadata}

\item{ubb}{UBB TRUE or FALSE}
}
\value{
data frame (Headers)
}
\description{
Returns text for reports (headers)
}
