% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{export_plotHTML}
\alias{export_plotHTML}
\title{Export a plot HTML}
\usage{
export_plotHTML(meta, snr, audience, report, data, ubb, export = TRUE)
}
\arguments{
\item{meta}{Meta data}

\item{snr}{School number}

\item{audience}{Audience}

\item{report}{Report template}

\item{data}{Data}

\item{ubb}{UBB}

\item{export}{Export}
}
\description{
Function creates and export a plot.
}
