% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{export_tables}
\alias{export_tables}
\title{Export tables}
\usage{
export_tables(meta, snr, data, audience, report, ubb)
}
\arguments{
\item{meta}{Meta data}

\item{snr}{SNR}

\item{data}{Data}

\item{audience}{Reporting group}

\item{report}{Reporttemplate}

\item{ubb}{UBB TRUE or FALSE}
}
\value{
A flextable
}
\description{
Export all plots as svg and convert them to PDFs
}
