% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_directory}
\alias{get_directory}
\title{Get the directory}
\usage{
get_directory(snr)
}
\arguments{
\item{snr}{School number}
}
\value{
String
}
\description{
Returns the main directory of the results
}
