% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_directory_res}
\alias{get_directory_res}
\title{Get the directory of the results}
\usage{
get_directory_res(snr, audience)
}
\arguments{
\item{snr}{School number}

\item{audience}{Report audience}
}
\value{
String
}
\description{
Get the sub-directory of the results
}
