% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_n}
\alias{get_n}
\title{Get the number of respondents from Limesurvey}
\usage{
get_n(audience, data)
}
\arguments{
\item{audience}{For which audience}

\item{data}{Data frame with survey data}
}
\description{
Get the number of respondents from Limesurvey
}
