% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_parameter}
\alias{get_parameter}
\title{Get Parameters}
\usage{
get_parameter(snr, audience, ubb, stype, ganztag)
}
\arguments{
\item{snr}{School number}

\item{audience}{Audience}

\item{ubb}{UBB}

\item{stype}{School type}

\item{ganztag}{Ganztag}
}
\value{
List of parameters
}
\description{
Get all survey parameter (reportpackage, valid n, etc,)
and raw data from Lime Survey
}
