% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_rid_of_blankspace}
\alias{get_rid_of_blankspace}
\title{Get Rid of Blankspace}
\usage{
get_rid_of_blankspace()
}
\value{
Alert.
}
\description{
Helper function to get rid of blankspace. Run steps manually.
}
