% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_rmd}
\alias{get_rmd}
\title{Get the rmd code for the report}
\usage{
get_rmd(x_seq)
}
\arguments{
\item{x_seq}{Sequence}
}
\value{
Returns a plot
}
\description{
Functions runs export_plots function for testing a  single plot
}
