% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_rprtpckg}
\alias{get_rprtpckg}
\title{Get report package (report package, survey, and report template)}
\usage{
get_rprtpckg(report, school, ubbs, allday)
}
\arguments{
\item{report}{Which report type}

\item{school}{Which school type}

\item{ubbs}{UBB TRUE or FALSE}

\item{allday}{Ganztag TRUE or FALSE}
}
\description{
Get rprtpckg (and survey) for a school
}
