% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_sname}
\alias{get_sname}
\title{Get school name}
\usage{
get_sname(snr)
}
\arguments{
\item{snr}{School number as a character string}
}
\description{
Get school name (string) based on their school number
}
