% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_snrlist}
\alias{get_snrlist}
\title{Get a List with School Numbers}
\usage{
get_snrlist(drop = TRUE, append = FALSE)
}
\arguments{
\item{drop}{Drop survey with no completed response}

\item{append}{Append Overall report}
}
\description{
Get school name (string) based on their school number
}
