% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{get_table}
\alias{get_table}
\title{Get table}
\usage{
get_table(meta, snr, audience, report, data, export = FALSE, ubb)
}
\arguments{
\item{meta}{Metadata}

\item{snr}{SNR}

\item{audience}{Audience of the report}

\item{report}{Reporttemplate}

\item{data}{Data}

\item{export}{Export}

\item{ubb}{UBB TRUE or FALSE}
}
\value{
A flextable
}
\description{
Get a flextable for RAW data input
}
