% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{header_reports}
\alias{header_reports}
\title{The headers reports data with all headers}
\format{
A data frame with 141 rows and 6 variables:
\describe{
\item{sort}{The sorting}
\item{plot}{The plot name}
\item{header1}{The first header}
\item{header2}{The second header}
\item{timestamp}{Timestamp of creation}
\item{report}{Report type}
...
}
}
\description{
This data which headers are displayed by each plot of the survey.
}
