% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{master_to_template}
\alias{master_to_template}
\title{The Master to Templates data}
\format{
A data frame with 104 rows and 9 variables:
\describe{
\item{sart}{Schulart}
\item{pckg}{Packet}
\item{survey}{Indicator for survey or not}
\item{surveyls_title}{Survey title in Lime Survey}
\item{template}{Template name}
\item{rpt}{Report}
\item{rpt_overall}{Report Overall}
\item{surveyID}{Survey ID Lime Survey}
\item{bfr_grp}{Befragtengruppe}
\item{timestamp}{Timestamp}
...
}
}
\description{
This data contains the school names.
}
