% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{plotGetData}
\alias{plotGetData}
\title{Daten fuer Grafik aus Gesamtdatensatz abfragen}
\usage{
plotGetData(data, rprtpckg, report, plotid, audience)
}
\arguments{
\item{data}{Datensatz aus surveyGetDataLongformat()}

\item{rprtpckg}{Reportpackage}

\item{report}{Report}

\item{plotid}{PlotID from meta data}

\item{audience}{Audience of the report}
}
\value{
data.frame
}
\description{
Daten fuer Grafik aus Gesamtdatensatz abfragen
}
\details{
Fragt aus Gesamtdatensatz (surveyGetDataLongformat()) anhand Schulart und Plotid
Daten fuer eine Grafik ab.
}
