% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{plotGetMetaData}
\alias{plotGetMetaData}
\title{Abrufen der Metadaten fuer ein Befragungspaket}
\usage{
plotGetMetaData(rprtpckg, report, audience, ganztag)
}
\arguments{
\item{rprtpckg}{(string)}

\item{report}{(string)}

\item{audience}{Group indicator}

\item{ganztag}{Ganztag indicator}
}
\value{
meta (vector string)
}
\description{
Abrufen der Metadaten fuer ein Befragungspaket
}
\details{
Ruft fuer einen Befragungspaketnamen eine Liste ab, die das Befragungspaket
und die plotid enthaellt.
}
