% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{plots_headers}
\alias{plots_headers}
\title{The plot headers data}
\format{
A data frame with 112 rows and 5 variables:
\describe{
\item{sort}{The sorting}
\item{plot}{The plot name}
\item{header1}{The first header}
\item{header2}{The second header}
\item{timestamp}{Timestamp of creation}
...
}
}
\description{
This data which headers are displayed by each plot of the survey.
}
