% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{plots_headers_ubb}
\alias{plots_headers_ubb}
\title{The UBB plot headers}
\format{
A data frame with 112 rows and 5 variables:
\describe{
\item{sort}{The sorting}
\item{plot}{The plot name}
\item{header1}{The first header}
\item{timestamp}{Timestamp of creation}
...
}
}
\description{
This data defines sets, lables, colors, etc.
}
