% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ReportPreview}
\alias{ReportPreview}
\title{Get a Preview of the Report}
\usage{
ReportPreview(snr, audience, ubb, ganztag, stype)
}
\arguments{
\item{snr}{The school number.}

\item{audience}{The audience for the report.}

\item{ubb}{Whether the school is a UBB school.}

\item{ganztag}{Whether the school is a Ganztagsschule.}

\item{stype}{The school type.}
}
\value{
Shiny app
}
\description{
This function invokes the Shiny app to preview the report
generated by the ReportMaster package.
}
\examples{
\dontrun{
ReportPreview(snr = "8934", audience = "sus", ubb = FALSE, ganztag = FALSE, stype = "gm")
}
}
