% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{reports}
\alias{reports}
\title{The Report Data}
\format{
A data frame with 112 rows and 5 variables:
\describe{
\item{report}{The Report name}
\item{vars}{The variable name}
\item{plot}{The plot name}
\item{label}{The label}
\item{label_short}{The label short}
\item{sets}{The sets}
\item{type}{The type}
\item{timestamp}{Timestamp of creation}
...
}
}
\description{
This data defines sets, lables, colors, etc.
}
