% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{run_Parallel}
\alias{run_Parallel}
\title{Run in Parallel Mode}
\usage{
run_Parallel(snr, audience, ubb, results, ganztag, stype)
}
\arguments{
\item{snr}{School number}

\item{audience}{Audience}

\item{ubb}{UBB}

\item{results}{Results: Text string for Audience}

\item{ganztag}{Ganztagsschule}

\item{stype}{School type}
}
\value{
Returns PDF file
}
\description{
Run all functions to create a report in parallel
}
