% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{runParallel}
\alias{runParallel}
\title{Run Parallel}
\usage{
runParallel(snr, audience, ubb, results, ganztag, stype)
}
\arguments{
\item{snr}{Schoolnumber}

\item{audience}{Report audience}

\item{ubb}{UBB TRUE or FALSE}

\item{results}{Results: Text string for audience of the report}

\item{ganztag}{Ganztagsschule TRUE or FALSE}

\item{stype}{Schooltype}
}
\value{
Returns Rds, Plots, and PDF report
}
\description{
Run all functions to create a report parallel
}
