% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{sets}
\alias{sets}
\title{Sets data}
\format{
A data frame with 54 rows and 7 variables:
\describe{
\item{set}{The set name}
\item{code}{The numerical value code}
\item{labels}{The labels}
\item{sort}{The sorting}
\item{colors}{The colors}
\item{text_color}{The text color}
\item{timestamp}{Timestamp of creation}
...
}
}
\description{
This data defines sets, lables, colors, etc.
}
