% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyConnectLs}
\alias{surveyConnectLs}
\title{Connect with Limesurvey}
\usage{
surveyConnectLs(server, user, credential)
}
\arguments{
\item{server}{Sever name}

\item{user}{Limesurvey user}

\item{credential}{Limesurvey credential}
}
\value{
session id aus Limesurvey
}
\description{
Verbindung mit Limesurvey aufbauen
}
