% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetData}
\alias{surveyGetData}
\title{Get Survey Data}
\usage{
surveyGetData(df, response = "short")
}
\arguments{
\item{df}{Vector with Ids}

\item{response}{Response indicator for Lime Survey}
}
\value{
data frame
}
\description{
Queries complete data records from Limesurvey for a list
of survey IDs.
}
