% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetDataLongformat}
\alias{surveyGetDataLongformat}
\title{Get survey data in long format}
\usage{
surveyGetDataLongformat(ids, ubb)
}
\arguments{
\item{ids}{Vector mit Survey-Ids aus Limesurvey}

\item{ubb}{TRUE or FALSE}
}
\value{
data.frame mit Daten mehrerer Surveys im Langformat
}
\description{
Baut aus Metadaten, Surveydatensatz und Variablenlabels mehrerer Surveys einen Datensatz.
}
\details{
Baut aus einer Liste von Survey-Ids aus den Metadaten, Surveydaten und Variablenlabels mehrerer Surveys einen Datensatz im Langformat.
Nutz dafuer die Funktionen: surveyGetMetaData(); surveyGetData() und surveyGetVariableLabels()
}
