% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetMetaData}
\alias{surveyGetMetaData}
\title{Metadaten von Surveys abrufen}
\usage{
surveyGetMetaData(ids)
}
\arguments{
\item{ids}{Vector mit Ids}
}
\value{
data.frame mit Metadaten der Surveys
}
\description{
Fragt fuer eine Liste von Survey-Ids aus Limesurvey die Metadaten
der Befragungen ab
}
