% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetResponseStat}
\alias{surveyGetResponseStat}
\title{Get Completed Responses of a Survey}
\usage{
surveyGetResponseStat(id)
}
\arguments{
\item{id}{Survey ID}
}
\value{
data frame
}
\description{
Queries the number of completed responses with get_summary method.
}
