% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetSurveyIds}
\alias{surveyGetSurveyIds}
\title{Gets metadata of the surveys}
\usage{
surveyGetSurveyIds(snr, year, ubb)
}
\arguments{
\item{snr}{Schoolnumber}

\item{year}{Year}

\item{ubb}{ubb}
}
\value{
vector mit IDs der Surveys einer Schule
}
\description{
Fragt fuer eine Liste von Survey-Ids aus Limesurvey die Metadaten
der Befragungen ab.
}
