% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetVariableLabels}
\alias{surveyGetVariableLabels}
\title{Variablenname und -Label mehrerer Surveys aus Limesurvey abrufen}
\usage{
surveyGetVariableLabels(tmp.ques.df)
}
\arguments{
\item{tmp.ques.df}{data.frame mit Metadaten der Befragungen (surveyGetMetaData())}
}
\value{
data.frame mit Variablennamen und -Labels mehrerer Surveys
}
\description{
Variablenname und -Label mehrerer Surveys aus Limesurvey abrufen
}
\details{
Ruft fuer eine Liste von Survey-Ids Variablennamen und Variablenlabels ab.
Nutz dafuer die Funktion surveyGetVariableRawLabels.
}
