% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetVariableRawColumnnames}
\alias{surveyGetVariableRawColumnnames}
\title{Variablenlabels aus Limesurvey in Base64 abrufen}
\usage{
surveyGetVariableRawColumnnames(iSurveyID, sHeadingType = "full")
}
\arguments{
\item{iSurveyID}{Survey-Id}

\item{sHeadingType}{full = Variablenlabel (Frage) | code = Variablenname (Code)}
}
\value{
vector mit Variablennamen bzw. Variablenlabels
}
\description{
Variablenlabels aus Limesurvey in Base64 abrufen
}
\details{
Ruft fuer eine Survey-Id den Surveydatensatz ab und gibt die erste Zeile mit den
Variablenlabels als Vector zurueck.
}
