% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetVariableRawLabels}
\alias{surveyGetVariableRawLabels}
\title{Variablenname und -Label aus Limesurvey abrufen}
\usage{
surveyGetVariableRawLabels(iSurveyID)
}
\arguments{
\item{iSurveyID}{Survey-Id}
}
\value{
data.frame mit Variablennamen und -Labels
}
\description{
Variablenname und -Label aus Limesurvey abrufen
}
\details{
Ruft fuer eine Survey-Id Variablennamen und Variablenlabels ab.
Nutz dafuer die Funktion surveyGetVariableRawColumnnames().
}
