% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{TableWorldCloud}
\alias{TableWorldCloud}
\title{Table for the Word Cloud}
\usage{
TableWorldCloud(data)
}
\arguments{
\item{data}{Data}
}
\value{
A flextable table
}
\description{
Get a table for the Word Cloud Text
}
