% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{templates}
\alias{templates}
\title{The Templates data}
\format{
A data frame with 112 rows and 5 variables:
\describe{
\item{stype}{The school type}
\item{ganztag}{Ganztag}
\item{type}{Survey type}
\item{surveys}{Survey template}
\item{rprtpckg}{Report package}
\item{ubb}{UBB}
\item{report_tmpl}{Report template}
\item{timestamp}{Timestamp of creation}
...
}
}
\description{
This data defines muster, survey, and report template.
}
