% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{test_ExportFun}
\alias{test_ExportFun}
\title{Test Export Function}
\usage{
test_ExportFun(testmeta)
}
\arguments{
\item{testmeta}{Test meta data}
}
\value{
Returns a plot
}
\description{
Functions runs export_plots function for testing a single plot
}
