% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{test_ExportPlot}
\alias{test_ExportPlot}
\title{Test Export Plot Function}
\usage{
test_ExportPlot(testmeta)
}
\arguments{
\item{testmeta}{Testmeta}
}
\value{
Returns a plot
}
\description{
Functions runs export_plots function for testing a  single plot
}
