% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_response}
\alias{check_response}
\title{Check whether a survey has valid response}
\usage{
check_response(snr, audience, ubb)
}
\arguments{
\item{snr}{School number}

\item{audience}{Audience}

\item{ubb}{UBB TRUE or FALSE}
}
\description{
Check if a survey has valid response
}
