% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{create_allplots2}
\alias{create_allplots2}
\title{Create all report plots at once simplified}
\usage{
create_allplots2(meta, snr, audience, report, data, ubb)
}
\arguments{
\item{meta}{Meta data}

\item{snr}{School number}

\item{audience}{Audience}

\item{report}{Report template}

\item{data}{Data}

\item{ubb}{UBB}
}
\description{
Create all report plots at once simplified (direct export)
}
