% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_directories}
\alias{create_directories}
\title{Create directories}
\usage{
create_directories(snr, audience, ubb)
}
\arguments{
\item{snr}{School number}

\item{audience}{Audience}

\item{ubb}{UBB}
}
\description{
Create directories for the report
}
