% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debugme}
\alias{debugme}
\title{Debug the Code}
\usage{
debugme(tmp.snr, tmp.ubb, tmp.ganztag, tmp.audience, tmp.stype, tmp.results)
}
\arguments{
\item{tmp.snr}{The snr}

\item{tmp.ubb}{The ubb}

\item{tmp.ganztag}{The ganztag}

\item{tmp.audience}{The audience}

\item{tmp.stype}{The stype}

\item{tmp.results}{The results string}
}
\value{
Error message
}
\description{
Function runs the code and returns error messages.
}
