% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{export_headers}
\alias{export_headers}
\title{Export table headers (Superseded by Parallelization)}
\usage{
export_headers(meta, ubb)
}
\arguments{
\item{meta}{Meta data}

\item{ubb}{UBB}
}
\value{
data
}
\description{
Returns text (headers) for the reports
}
