% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{export_plot}
\alias{export_plot}
\title{Export plots}
\usage{
export_plot(meta, snr, audience, report, data, ubb, export = TRUE)
}
\arguments{
\item{meta}{Meta data}

\item{snr}{Schoolnumber}

\item{audience}{audience}

\item{report}{Reporttemplate}

\item{data}{Data}

\item{ubb}{TRUE or FALSE}

\item{export}{TRUE or FALSE}
}
\description{
Function to export pngs directly.
}
