% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{export_tables}
\alias{export_tables}
\title{Export tables}
\usage{
export_tables(meta, snr, data, audience, report, ubb)
}
\arguments{
\item{meta}{Meta data}

\item{snr}{School number}

\item{data}{Data}

\item{audience}{Audience}

\item{report}{Report template}

\item{ubb}{UBB}
}
\value{
A flextable table
}
\description{
Export all tables as svg and convert them to pdf
}
