% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{generate_rmd}
\alias{generate_rmd}
\title{Generate the Rmd file for the report}
\usage{
generate_rmd(x_seq, ubb, export = TRUE, file_name = "generated_document.Rmd")
}
\arguments{
\item{x_seq}{Sequence}

\item{ubb}{UBB TRUE or FALSE}

\item{export}{Export TRUE or FALSE}

\item{file_name}{Filename}
}
\value{
Returns a plot
}
\description{
Functions runs export_plots function for testing a  single plot
}
