% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_directory_res}
\alias{get_directory_res}
\title{Get the directory of the results}
\usage{
get_directory_res(snr, audience)
}
\arguments{
\item{snr}{The schoolnumber}

\item{audience}{The audience of the report}
}
\value{
Foldername
}
\description{
Bla bla
}
