% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_n}
\alias{get_n}
\title{Get the Number of Respondents from Lime Survey}
\usage{
get_n(audience, data)
}
\arguments{
\item{audience}{Audience}

\item{data}{Data frame with survey data}
}
\description{
Get the Number of Respondents from Lime Survey
}
