% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_parameter}
\alias{get_parameter}
\title{Get parameters}
\usage{
get_parameter(snr, audience, ubb, stype, ganztag)
}
\arguments{
\item{snr}{Schoolnumber}

\item{audience}{Report audience}

\item{ubb}{UBB TRUE or FALSE}

\item{stype}{School type}

\item{ganztag}{Ganztag TRUE or FALSE}
}
\value{
List of parameters
}
\description{
Get all survey parameter (reportpackage, valid n, etc,)
and raw data from limesurvey
}
