% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_rmd}
\alias{get_rmd}
\title{Get Rmd Code for the Report}
\usage{
get_rmd(x_seq)
}
\arguments{
\item{x_seq}{Sequence}
}
\value{
Character
}
\description{
Depending on the number of plots and tables, the function
creates the Rmd chunks for the report
}
