% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_rprtpckg}
\alias{get_rprtpckg}
\title{Get Report Package (report package, survey, and report template)}
\usage{
get_rprtpckg(report, school, ubbs, allday)
}
\arguments{
\item{report}{Report template}

\item{school}{School type}

\item{ubbs}{UBB}

\item{allday}{Ganztag}
}
\description{
Get rprtpckg (and survey) for a school
}
