% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_snrlist}
\alias{get_snrlist}
\title{Get a list with school numbers}
\usage{
get_snrlist(drop = TRUE, append = FALSE)
}
\arguments{
\item{drop}{Drop TRUE or FALSE}

\item{append}{Append TRUE or FALSE}
}
\description{
Get school name (string) based on their school number
}
