% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{get_table}
\alias{get_table}
\title{Get a Table of the Report}
\usage{
get_table(meta, snr, audience, report, data, export = FALSE, ubb)
}
\arguments{
\item{meta}{Metadata}

\item{snr}{School number}

\item{audience}{Audience}

\item{report}{Report template}

\item{data}{Data}

\item{export}{Export}

\item{ubb}{UBB}
}
\value{
A flextable table
}
\description{
Get a table of the report for each plot
}
