% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{schools_names}
\alias{schools_names}
\title{The school names data.}
\format{
A data frame with 54 rows and 7 variables:
\describe{
\item{SNR}{The school number}
\item{SNAME}{School name}
...
}
}
\description{
This data contains the school names.
}
