% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyConnectLs}
\alias{surveyConnectLs}
\title{Connect with Limesurvey}
\usage{
surveyConnectLs(server, user, credential)
}
\arguments{
\item{server}{Sever name}

\item{user}{Lime Survey user}

\item{credential}{Lime Survey credential}
}
\value{
Sesssion ID
}
\description{
Creates the connection to Limesurvey
}
