% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetData}
\alias{surveyGetData}
\title{Surveys abrufen}
\usage{
surveyGetData(df, response = "short")
}
\arguments{
\item{df}{Vector mit Ids}

\item{response}{Response indicator for Limesurvey}
}
\value{
data.frame
}
\description{
Fragt fuer eine Liste von Survey-Ids vollstaendige Datensaetze
aus Limesurvey ab.
}
