% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetMetaData}
\alias{surveyGetMetaData}
\title{Get Meta Data of the Surveys}
\usage{
surveyGetMetaData(ids)
}
\arguments{
\item{ids}{Survey IDs}
}
\value{
Data frame
}
\description{
Queries the metadata of the surveys for a list of survey IDs
from Limesurvey
}
