% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetResponseStat}
\alias{surveyGetResponseStat}
\title{Anzahl full responses eines Surveys abrufen}
\usage{
surveyGetResponseStat(id)
}
\arguments{
\item{id}{id (num)}
}
\value{
df
}
\description{
Fragt mit get_summary method die Anzahl der full responses ab.
}
