% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{surveyGetSurveyIds}
\alias{surveyGetSurveyIds}
\title{Gets Meta Data of the Surveys}
\usage{
surveyGetSurveyIds(snr, year, ubb)
}
\arguments{
\item{snr}{School number}

\item{year}{Year}

\item{ubb}{UBB}
}
\value{
Vector with survey IDs
}
\description{
Queries the survey's metadata for a list of survey IDs
from Lime Survey
}
